+++ src/message.h
@@ -34,6 +34,6 @@
 int32_t prepare_message(char *header, char *data, Message *message);
 int32_t print_message(Message *message);
 
-int32_t message_to_bytes(Message * message, unsigned char * byte_array_result, int32_t * byte_array_size);
+int32_t message_to_bytes(Message * message, unsigned char * byte_array_result, size_t * byte_array_size);
 int32_t message_bytes_to_message(unsigned char * avro_byte_stream, int32_t bytesToRead, Message * message);
 #endif //SERVER_CLIENT_MESSAGE_H
